Move access control into the right place
